Fix congested topic issue
This was caused due to the large difference in processing speed between the video source producer and some of the stream apps.

What was happening was the producer was finishing, thus sending its tombstone message downstream.
The slow object & pose detection processors were still processing frames from earlier.
But, because there is an aggregate consumer subscribed both to the source topic _And_ their output topics, it was exiting early,
leaving no consumers to read from their output topics.

This was causing the slow processors to fill up their output topics' buffers, leading to a deadlock

The current solution is for the aggregate consumer to wait for all its downstream consumers, rather than failing fast